refactor: prepare for docs-content changes (#837)
We recently refactored how our docs-content example module is composed.
We did this because the docs app assumes path in the docs-content
package that are not necessarily always correct. i.e. breaking the chips
drag and drop example.

Additionally the docs-content example module is now more self-contained
so that assumptions about files inside of an example are not required
anymore. i.e. examples can now _not_ have CSS or HTML files. Previously
we always had to display an empty CSS/HTML file (which is confusing).