build: re-add ngcc postinstall for docs content (#864)
We recently removed the ngcc postinstall since the Angular
compiler webpack plugin dynamcially processed packages that
were imported. This reduced Yarn install time for the docs site.

Now due to a recent update of the CLI build-angular version,
the examples break because ngcc no longer processes entry-points
based on module dependencies, but rather only as per TypeScript
program source files/referenced files. This does not include the
docs-content as the examples are loaded lazily and aren't part of
the TS compilation.