Fix forking and saving for webcontainer stackblitz examples
Fixes the forking and saving functionality for WebContainer StackBlitz
examples. The saving/forking functionality did not work (no proper error though)
because we delivered a binary `.ico` file to StackBlitz.

To fix the examples, we will remove the `favicon.ico` file, even the goal
was to keep the examples as equivalent as possible to a `ng new`-created app.