ci: enable `ASPECT_RULES_JS_FROZEN_PNPM_LOCK` to ensure outdated pnpm file is flagged
In the migration period, until we decide to switch to `pnpm` fully, a pnpm lock file needs to be kept in sync. By default, `rules_js` will try to sync it automatically, but rather we would want to flag it as an error on CI if the pnpm file is outdated- so that not all users need to pay the cost of having to run `pnpm import` from time to time.